// src/app/models/iTutorial.ts
export interface ITutorial {
    title: string;
    date: string;
    order: number;
    category: string;
    categoryImage: string;
    mainImage: string;
    authorImage: string;
    authorName: string;
    authorRole: string;
    slug: string;
    description: string;
    draft: boolean;
}